The component has not use as of now but will be used later, moved both userListItems into a folder
