+++ src/components/userListComponents/userListItemSpecialism/userListItemSpecialism.css
